Gulpified the project structure.
Gulp task now runs build, nodemon, and browser sync.
Project shows basic layout.
All build, watching and syncing works *I think*
