+++ client/app/app.html
@@ -3,59 +3,6 @@
 	<h1>Welcome to FlyptoX</h1>
 	</head>
 
-  <style type="text/css">
-    body {
-      font-family: arial, sans-serif;
-      background: #EEEEEE;
-    }
-    body h1 {
-      font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-      font-size: 3.5em;
-      color: gray;
-      text-align: center;
-    }
-    body h2,
-    body h3,
-    body h4,
-    body h5,
-    body h6 {
-      font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-      font-family: Helvetica Neue;
-      font-size: 14px;
-      font-style: normal;
-      font-variant: normal;
-      font-weight: 500;
-      line-height: 15.3999996185303px;
-      margin-top: 5px;
-    }
-    body .order {
-      background: #ff7b81;
-      height: 30px;
-      width: 100px;
-    }
-    body .tx {
-      background: #ffffff;
-      height: 870px;
-      width: 300px;
-      position: absolute;
-    }
-    body .execution {
-      background: #ffffff;
-      height: 200px;
-      width: 300px;
-    }
-    body .close {
-      background: #4b6e84;
-      height: 30px;
-      width: 300px;
-    }
-    .auth {
-      background: #4b6e84;
-      height: 30px;
-      width: 60px;
-      position: absolute;
-    }
-  </style>
 <body>
 
 <div ng-controller="signUpCtrl">
