--- client/app/components/chart/techan/chart.css
@@ -1,28 +1,0 @@
-/*body {
-  font: 10px sans-serif;
-}
-
-.axis path,
-.axis line {
-  fill: none;
-  stroke: #000;
-  shape-rendering: crispEdges;
-}
-
-path.candle {
-  stroke: #000000;
-}
-
-path.candle.body {
-  stroke-width: 0;
-}
-
-path.candle.up {
-  fill: #00AA00;
-  stroke: #00AA00;
-}
-
-path.candle.down {
-  fill: #FF0000;
-  stroke: #FF0000;
-}*/
