--- client/app/components/chart/techan/chart.html
@@ -1,14 +1,0 @@
-<!-- <!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title></title>
-    <link rel="stylesheet" type="text/css" href="chart.css">
-  </head>
-  <body>
-    <button>Reset</button>
-    <script src="http://d3js.org/d3.v3.min.js"></script>
-    <script src="http://techanjs.org/techan.min.js"></script>
-    <script src="chart.js"></script>
-  </body>
-</html> -->
