--- client/app/components/orderbook/orderbook.js
@@ -1,13 +1,0 @@
-// FlyptoX.controller('orderbookCtrl', function () {
-//   	// $scope.data = {};
-//   	// 	orderSize: 10,
-//   	// 	price: 50,
-//   	// 	filled: "yes"
-//   	// };
-//   });
-
-// var myApp = angular.module('myApp',[]);
-
-// myApp.controller('GreetingController', ['$scope', function($scope) {
-//   $scope.greeting = 'Hola!';
-// }]);
