+++ server/views/index.html
@@ -2,17 +2,15 @@
 <html lang="en">
     <head>
         <base href="/">
-<!--         <link href="/assets/css/app.less" rel="stylesheet" /> -->
+        <link href="/assets/css/main.min.css" rel="stylesheet" />
 
         <!-- build:js /app.js -->
 
         <!-- bower -->
-        <script src="/lib/angular/angular.js"></script>
         <!-- endbower -->
 
         <!-- appcode -->
-        <script src="/app/modules.js"></script>
-        <script src="/app/app.js"></script>
+        <script src="/app/main.min.js"></script>
         <!-- endappcode -->
 
         <!-- endbuild -->
